requirements.txt
requests~=2.28.1
beautifulsoup4~=4.11.1
openpyxl~=3.0.10
xmltodict~=0.13.0
Flask~=2.2.2
argparse
dnspython
times
asyncio
aiohttp
